fix(travis): prevent double builds on PRs
Without this, if a PR is opened for a branch, Travis creates two builds;
one for the branch and one for the PR itself. When Travis is configured
(in the UI) to build both the push and PR (the default), I think a
better default for builds is to trigger only _one_ build.

Semantic release itself uses this set up, see:
https://github.com/semantic-release/commit-analyzer/issues/11